commit cuoi ngay , cap nhat avatar fragment account , voi fragment update thanhvien
