+++ app/src/main/java/com/example/myapplication/model/DAO/ThanhVienDAO.java
@@ -31,22 +31,11 @@ public interface ThanhVienDAO {
             "OR email = :email OR so_dien_thoai = :soDienThoai)")
     boolean isThanhVienExist(String tenDangNhap, String email, String soDienThoai);
 
-    @Query("SELECT mat_khau FROM table_thanhvien WHERE ten_dang_nhap = :tenDangNhap")
-    String getMatKhauByUserName(String tenDangNhap);
-
     @Query("SELECT id_quyen FROM table_thanhvien WHERE ten_dang_nhap = :tenDangNhap")
     int getQuyenByUserName(String tenDangNhap);
 
     @Query("SELECT * FROM table_thanhvien WHERE email = :inputEmail LIMIT 1")
     ThanhVien getThanhVienByEmail(String inputEmail);
-
-    @Query("UPDATE table_thanhvien SET mat_khau = :password WHERE email = :email")
-    int updatePasswordByEmail(String password, String email);
-
     @Query("SELECT * FROM table_thanhvien WHERE ten_dang_nhap = :username LIMIT 1")
     ThanhVien getThanhVienByUserName(String username);
-
-    @Query("UPDATE table_thanhvien SET ten_thanh_vien = :ten, ho_thanh_vien = :ho, so_dien_thoai = :soDienThoai, email = :email, ngay_tao = :ngayTao, ngay_cap_nhat = :ngayCapNhat, avatar = :avatar WHERE ten_dang_nhap = :tenDangNhap")
-    void updateByUsername(String tenDangNhap, String ho, String ten, String soDienThoai, String email, String ngayTao, String ngayCapNhat, String avatar);
-
 }
