+++ app/src/main/java/com/example/myapplication/view/UpdateThanhVienFragment.java
@@ -17,12 +17,16 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
+import android.widget.Toast;
 
+import com.bumptech.glide.Glide;
 import com.example.myapplication.R;
 import com.example.myapplication.config.FunctionPublic;
 import com.example.myapplication.databinding.FragmentUpdateThanhVienBinding;
 import com.example.myapplication.model.ThanhVien;
 import com.example.myapplication.viewmodel.UpdateThanhVienViewModel;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
 
 
 public class UpdateThanhVienFragment extends Fragment {
@@ -81,7 +85,7 @@ public class UpdateThanhVienFragment extends Fragment {
                 result -> {
                     if (result.getResultCode() == RESULT_OK && result.getData() != null) {
                         Uri selectedImageUri = result.getData().getData();
-                        updateThanhVienViewModel.uploadImageToFireBase(selectedImageUri, getContext());
+                        uploadImageToFireBase(selectedImageUri);
                     }
                 }
         );
@@ -97,6 +101,45 @@ public class UpdateThanhVienFragment extends Fragment {
         });
         // Inflate the layout for this fragment
         return fragmentUpdateThanhVienBinding.getRoot();
+    }
+
+    public void uploadImageToFireBase(Uri uri){
+        // Tham chiếu đến Firebase Storage
+        StorageReference storageRef = FirebaseStorage.getInstance().getReference();
+
+        // Tạo tên file duy nhất cho hình ảnh (ví dụ: sử dụng thời gian hiện tại làm tên file)
+        String fileName = System.currentTimeMillis() + ".jpg";
+
+        // Tạo tham chiếu đến file trên Firebase Storage
+        StorageReference imageRef = storageRef.child("images/" + fileName);
+
+        // Upload hình ảnh lên Firebase Storage
+        imageRef.putFile(uri)
+                .addOnProgressListener(snapshot -> {
+                    Toast.makeText(getContext(), "Đang tải ảnh ", Toast.LENGTH_LONG).show();
+                })
+                .addOnSuccessListener(taskSnapshot -> {
+                    // Upload thành công, tiến hành hiển thị hình ảnh
+                    displayImageFromFirebaseStorage(imageRef);
+                    Toast.makeText(getContext(), "Tải ảnh thành công ", Toast.LENGTH_LONG).show();
+                })
+                .addOnFailureListener(e -> {
+                    // Upload thất bại, xử lý lỗi tại đây (nếu cần)
+                });
+        // Gọi phương thức xử lý ảnh được chọn
+
+    }
+
+    private void displayImageFromFirebaseStorage(StorageReference imageRef) {
+        // Lấy URL của hình ảnh đã tải lên từ Firebase Storage
+        imageRef.getDownloadUrl()
+                .addOnSuccessListener(uri -> {
+                    updateThanhVienViewModel.setAvatar(String.valueOf(uri));
+                    Glide.with(getContext()).load(updateThanhVienViewModel.getAvatar()).into(fragmentUpdateThanhVienBinding.avatarImg);
+                })
+                .addOnFailureListener(e -> {
+                    // Xử lý lỗi nếu không thể lấy URL hình ảnh (nếu cần)
+                });
     }
