+++ app/src/main/res/layout/fragment_account.xml
@@ -56,7 +56,7 @@
                         android:layout_height="wrap_content"
 
                         android:text="@={AccountViewModel.ho}"
-                        android:background="@color/white"
+                        android:background="#00FFFFFF"
                         android:hint="Họ "
                         android:paddingHorizontal="2dp"/>
                     <EditText
@@ -64,7 +64,7 @@
                         android:layout_height="wrap_content"
 
                         android:text="@={AccountViewModel.ten}"
-                        android:background="@color/white"
+                        android:background="#00FFFFFF"
                         android:paddingHorizontal="2dp"
                         android:hint="Tên "/>
                 </LinearLayout>
